Move bitcoin payment type preference from `PermissionService` to `KeyRingBitcoinService`
