+++ packages/background/src/permission/service.ts
@@ -15,7 +15,6 @@ import {
   ChainInfo,
   GENESIS_HASH_TO_NETWORK,
   GenesisHash,
-  SupportedPaymentType as BitcoinPaymentType,
 } from "@keplr-wallet/types";
 import { action, autorun, makeObservable, observable, runInAction } from "mobx";
 import { migrate } from "./migrate";
@@ -38,9 +37,6 @@ export class PermissionService {
   @observable
   protected currentBaseChainIdForBitcoinByOriginMap: Map<string, string> =
     new Map();
-
-  @observable
-  protected preferredBitcoinPaymentType: BitcoinPaymentType | undefined;
 
   constructor(
     protected readonly kvStore: KVStore,
@@ -131,16 +127,6 @@ export class PermissionService {
             }
           });
         }
-
-        const savedPreferredBitcoinPaymentType =
-          await this.kvStore.get<BitcoinPaymentType>(
-            "preferredBitcoinPaymentType/v1"
-          );
-        if (savedPreferredBitcoinPaymentType) {
-          runInAction(() => {
-            this.preferredBitcoinPaymentType = savedPreferredBitcoinPaymentType;
-          });
-        }
       }
     }
 
@@ -160,10 +146,6 @@ export class PermissionService {
       this.kvStore.set(
         "currentBaseChainIdForBitcoinByOriginMap/v1",
         Object.fromEntries(this.currentBaseChainIdForBitcoinByOriginMap)
-      );
-      this.kvStore.set(
-        "preferredBitcoinPaymentType/v1",
-        this.preferredBitcoinPaymentType
       );
     });
   }
@@ -914,26 +896,5 @@ export class PermissionService {
     } else {
       this.setCurrentBaseChainIdForBitcoin(origins, chainId);
     }
-  }
-
-  getPreferredBitcoinPaymentType(): BitcoinPaymentType {
-    return this.preferredBitcoinPaymentType ?? "taproot";
-  }
-
-  @action
-  setPreferredBitcoinPaymentType(paymentType: BitcoinPaymentType) {
-    // CHECK: origin permission 확인 필요 여부, keyring service로 옮겨야 할 지..
-    this.preferredBitcoinPaymentType = paymentType;
-
-    console.log("dispatchEvent keplr_bitcoinAccountsChanged...", paymentType);
-
-    this.interactionService.dispatchEvent(
-      WEBPAGE_PORT,
-      "keplr_bitcoinAccountsChanged",
-      {
-        paymentType,
-        origin: undefined,
-      }
-    );
   }
 }
